+++ client/src/types/index.d.ts
@@ -2,8 +2,8 @@ interface UserTypes {
   email: string;
   password: string;
   confirmPassword: string;
-  first_name: string;
-  last_name: string;
+  firstName: string;
+  lastName: string;
 }
 
 interface SvgTypes {
